+++ Tasks/UnityActivateLicense/UnityActivateLicenseV1/unity-activate-license.ts
@@ -2,38 +2,51 @@ import path = require('path');
 import tl = require('azure-pipelines-task-lib/task');
 import { getUnityEditorVersion } from './unity-build-shared';
 import {
-    UnityToolRunner,
     UnityPathTools,
     UnityVersionInfoResult,
     Utilities
 } from '@dinomite-studios/unity-azure-pipelines-tasks-lib';
 
-tl.setResourcePath(path.join(__dirname, 'task.json'));
-
-// Input variables.
-const usernameInputVariableName = 'username';
-const passwordInputVariableName = 'password';
+// Input variable names.
+export const usernameInputVariableName = 'username';
+export const passwordInputVariableName = 'password';
+export const versionInputVariableName = 'version';
+export const unityEditorsPathModeInputVariableName = 'unityEditorsPathMode';
+export const customUnityEditorsPathInputVariableName = 'customUnityEditorsPath';
+export const localPathInputVariableName = 'Build.Repository.LocalPath';
+export const versionSelectionModeVariableName = "versionSelectionMode";
 const serialInputVariableName = 'serial';
-const unityEditorsPathModeInputVariableName = 'unityEditorsPathMode';
-const customUnityEditorsPathInputVariableName = 'customUnityEditorsPath';
-const localPathInputVariableName = 'Build.Repository.LocalPath';
-const unityProjectPathInputVariableName = 'unityProjectPath';
-const unityVersionInputVariableName = 'unityVersion';
 
-/**
- * Main task runner. Executes the task and sets the result status for the task.
- */
-async function run() {
+function run() {
     try {
+        // Configure localization.
+        tl.setResourcePath(path.join(__dirname, 'task.json'));
+
         // Setup and read inputs.
         const username = tl.getInput(usernameInputVariableName, true)!;
         const password = tl.getInput(passwordInputVariableName, true)!;
         const serial = tl.getInput(serialInputVariableName, true)!;
-        const projectPath = tl.getPathInput(unityProjectPathInputVariableName) || '';
-        const unityVersion = { info: { version: tl.getInput(unityVersionInputVariableName) } } as UnityVersionInfoResult || getUnityEditorVersion();
+        const versionSelectionMode = tl.getInput(versionSelectionModeVariableName, true)!
         const unityEditorsPath = UnityPathTools.getUnityEditorsPath(
             tl.getInput(unityEditorsPathModeInputVariableName, true)!,
             tl.getInput(customUnityEditorsPathInputVariableName));
+
+        var unityVersion: UnityVersionInfoResult;
+        if (versionSelectionMode === 'specify') {
+            let customVersion = tl.getInput(versionInputVariableName, true)!;
+            unityVersion = {
+                info: {
+                    isAlpha: false,
+                    isBeta: false,
+                    version: customVersion,
+                    revision: undefined
+                },
+                error: undefined
+            }
+        } else {
+            unityVersion = getUnityEditorVersion();
+        }
+
         const unityExecutablePath = UnityPathTools.getUnityExecutableFullPath(unityEditorsPath, unityVersion.info!);
         const logFilesDirectory = path.join(tl.getVariable(localPathInputVariableName)!, 'Logs');
         const logFilePath = path.join(logFilesDirectory, `UnityActivationLog_${Utilities.getLogFileNameTimeStamp()}.log`);
@@ -49,12 +62,11 @@ async function run() {
             .arg('-username').arg(username)
             .arg('-password').arg(password)
             .arg('-serial ').arg(serial)
-            .arg('-projectPath').arg(projectPath)
             .arg('-logfile').arg(logFilePath);
-        const result = await UnityToolRunner.run(unityCmd, logFilePath);
+        const result = unityCmd.execSync();
 
         // Unity process has finished. Set task result.
-        if (result === 0) {
+        if (result.code === 0) {
             const activateLicenseSuccessLog = tl.loc('successLicenseActivated');
             console.log(activateLicenseSuccessLog);
             tl.setResult(tl.TaskResult.Succeeded, activateLicenseSuccessLog);