+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.1]
+
+### Added
+
+- New `Unity Setup` task (preview). Use this task to install a Unity editor and required modules before running your build
+
+### Changed
+
+- Updated `UnityActivateLicense` task to v1.2.0. Please see the task [CHANGELOG](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/blob/master/Tasks/UnityActivateLicense/UnityActivateLicenseV1/CHANGELOG.md) for mroe info
+- Updated `UnityBuild` to v3.2.1. Please seethe task [CHANGELOG](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/blob/master/Tasks/UnityBuild/UnityBuildV3/CHANGELOG.md) for more info
+
 ## [2.9.0]
 
 ### Changed