+++ Tasks/UnityBuild/UnityBuildV3/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.1]
+
+### Fixed
+
+- Fixed undefined path issue when specifying a custom unity editor version to build with
+
+### Changed
+
+- Updated dependencies
+
 ## [3.2.0]
 
 ### Fixed